+++ packages/apps/docs/src/components/Search/components/ResultCount.tsx
