+++ packages/apps/docs/src/components/Search/styles.ts
@@ -9,10 +9,23 @@ export const StaticResultsList: StyledComponent<'ul'> = styled('ul', {
   padding: 0,
 });
 
-export const ScrollBox: StyledComponent<'div'> = styled('div', {
-  height: '55vh',
-  overflowY: 'scroll',
+export const ScrollBox: StyledComponent<
+  'div',
+  { disabled?: 'true' | 'false' | boolean }
+> = styled('div', {
   margin: '$2 0',
+  defaultVariants: {
+    disabled: false,
+  },
+  variants: {
+    disabled: {
+      true: {},
+      false: {
+        height: '55vh',
+        overflowY: 'scroll',
+      },
+    },
+  },
 });
 
 export const StyledItem: StyledComponent<'a'> = styled('a', {
