+++ packages/apps/docs/src/hooks/useSearch/useSemanticSearch.ts
@@ -3,26 +3,35 @@ import { useEffect, useState } from 'react';
 interface ISemanticSearch {
   isLoading: boolean;
   results: ISearchResult[];
+  error?: string;
 }
+
+const searchErrorMessage =
+  'There was a problem. Sorry for the inconvenience. Please try again!';
 
 export const useSemanticSearch = (query?: string): ISemanticSearch => {
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [error, setError] = useState<boolean | undefined>(false);
+  const [error, setError] = useState<string | undefined>();
   const [results, setResults] = useState<ISearchResult[]>([]);
 
   const getSearchResults = async (query: string): Promise<void> => {
-    const response = await fetch(`/api/semanticsearch?query=${query}`);
+    try {
+      const response = await fetch(`/api/semanticsearch?query=${query}`);
 
-    const data = await response.json();
+      const data = await response.json();
 
-    if (data.status !== undefined && data.status >= 400) {
-      setError(error);
+      if (data.status !== undefined && data.status >= 400) {
+        setError(searchErrorMessage);
+        setResults([]);
+        return;
+      }
+
+      setError(undefined);
+      setResults(data);
+    } catch (e) {
+      setError(searchErrorMessage);
       setResults([]);
-      return;
     }
-
-    setError(undefined);
-    setResults(data);
   };
 
   useEffect(() => {
@@ -30,6 +39,7 @@ export const useSemanticSearch = (query?: string): ISemanticSearch => {
   }, [results]);
 
   useEffect(() => {
+    setError(undefined);
     if (query !== undefined) {
       setIsLoading(true);
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
@@ -40,5 +50,6 @@ export const useSemanticSearch = (query?: string): ISemanticSearch => {
   return {
     isLoading,
     results,
+    error,
   };
 };
